FinalProject_DetectiveGame/requirements.txt
# Core NLP library
spacy>=3.7.0

# Terminal UI/Colors
rich>=13.0.0

# Google API Usage
google-genai>=0.1.0